+++ src/torrent.cc
@@ -728,7 +728,7 @@ Torrent::callbackCompleteHashing(unsigned int piece, bool result)
 	/* At least someone has this piece... we! */
 	pieceCardinality[piece]++;
 	if (piece == numPieces - 1) {
-printf("final piece: left was %lu\n", left);
+printf("final piece: left was %llu\n", (unsigned long long) left);
 		left -= getTotalSize() % pieceLen > 0 ?
 		        getTotalSize() % pieceLen : pieceLen;
 printf("final piece: %lu, left=%lu\n",
